requirements.txt
# 医疗导诊系统依赖
# 核心框架
flask==2.3.3
streamlit==1.28.0

# AI相关
langchain==0.0.346
openai==0.28.0

# 数据验证和模型
pydantic==1.10.12

# 异步处理
aiohttp==3.8.5
asyncio==3.4.3
aiofiles==23.2.1

# 工具类
python-dotenv==1.0.0
requests==2.31.0

# 开发工具（可选）
black==23.7.0
flake8==6.0.0
mypy==1.5.1